ECA-1721: Only display an error message and not the full stacktrace when we try to remove an End Entity Profile that doesn't exist.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@10673 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/cesecore/core/ejb/ra/raadmin/EndEntityProfileSessionBean.java
@@ -495,18 +495,24 @@ public class EndEntityProfileSessionBean implements EndEntityProfileSessionLocal
      * Removes an end entity profile from the database.
      */
     public void removeEndEntityProfile(final Admin admin, final String profilename) {
+    	boolean success = false;
         try {
         	final EndEntityProfileData pdl = EndEntityProfileData.findByProfileName(entityManager, profilename);
-            entityManager.remove(pdl);
-            flushProfileCache();
-            final String msg = INTRES.getLocalizedMessage("ra.removedprofile", profilename);
-            logSession.log(admin, admin.getCaId(), LogConstants.MODULE_RA, new java.util.Date(), null, null,
-                    LogConstants.EVENT_INFO_ENDENTITYPROFILE, msg);
+        	if (pdl != null) {
+                entityManager.remove(pdl);
+                flushProfileCache();
+                final String msg = INTRES.getLocalizedMessage("ra.removedprofile", profilename);
+                logSession.log(admin, admin.getCaId(), LogConstants.MODULE_RA, new java.util.Date(), null, null,
+                        LogConstants.EVENT_INFO_ENDENTITYPROFILE, msg);
+                success = true;
+        	}
         } catch (Exception e) {
+            LOG.error("Error was caught when trying to remove end entity profile " + profilename, e);
+        }
+        if (!success) {
         	final String msg = INTRES.getLocalizedMessage("ra.errorremoveprofile", profilename);
             logSession.log(admin, admin.getCaId(), LogConstants.MODULE_RA, new java.util.Date(), null, null,
                     LogConstants.EVENT_ERROR_ENDENTITYPROFILE, msg);
-            LOG.error("Error was caught when trying to remove end entity profile " + profilename, e);
         }
     }
